med-scanner-py/requirements.txt
mysqlclient
django-cors-headers